+++ src/main/java/Application/Command/MoveCommand.java
@@ -1,4 +1,4 @@
-package Application;
+package Application.Command;
 
 import Domain.Board.Position;
