+++ src/main/java/Application/Command/SaveCommand.java
@@ -1,4 +1,4 @@
-package Application;
+package Application.Command;
 
 public record SaveCommand() implements GameCommand{
 }
